MFC r306443:
Fix an issue with accept_filter introduced with r261242:

As a side effect of r261242 when using accept_filter the
first call to soisconnected() is done earlier in tcp_input()
instead of tcp_do_segment() context.  Restore the expected behaviour.

Note:  This call to soisconnected() seems to be extraneous in all
cases (with or without accept_filter).  Will be addressed in a
separate commit.

PR:			212920
Reported by:		Alexey
Tested by:		Alexey, jch
Sponsored by:		Verisign, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@306922 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f