MFC r306740, r307150: groff: use changelog date in man pages
r306740:

groff: use changelog date rather than file modification date in man pages

The source checkout date is not particularly relevant, and this makes
groff man pages build reproducibly.

r307150:

Avoid using 'head' in generating groff doc date

It may not be available in certain cross build cases.

Note that this is a slight change in functionality, in that now only the
first line of the source ChangeLog file is processed. This is acceptable
as groff will be retired and we won't encounter a possibly-different
ChangeLog format.


git-svn-id: https://svn.freebsd.org/base/stable/11@307631 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f