MFC 289401: cxgbe(4): support for the kernel RSS option.
You need PCBGROUP and RSS in the kernel config to use this.

Note: Since RSS is not present in 10.x this is mostly a no-op and is
stubbed out by removing the #include of opt_rss.h.  This is merged
primarily to reduce conflicts in future merges, however it does add a
couple of diagnostic messages related to RSS buckets vs RX queue
counts.

Discussed with:	np
Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/stable/10@308153 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f