MFC r304316:
Rename kernel config A10 into ALLWINNER_UP as it is intend to work with all
Allwinner Uniprocessor SoC.
As of now it works with A10 and A13 (and possibly R8 as it is the same as the A13).
Move files.a10 into a1o subdirectory as it should be.
Rename std.a10 into std.allwinner_up


git-svn-id: https://svn.freebsd.org/base/stable/11@308273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f