MFC 277763,280146,287631: Various fixes to DDP.
277763:
Lock the socket buffer before jumping to the 'out' label if sblock()
fails in t4_soreceive_ddp().

280146:
Move special DDP handling for closing a connection into a new
handle_ddp_close() function in t4_ddp.c as the logic is similar
to handle_ddp_data().  This allows all knowledge of the special
DDP mbufs to be private to t4_ddp.c as well.

287631:
Add a comment to clarify how to determine the amount of received DDP
data.

Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/stable/10@308281 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f