Fix git tools when run against a worktree
In a git worktree, the gitdir is in an entirely different location.
In arcgit, use git rev-parse --git-dir to get the correct path to it
always.

When running git from outside of the work tree, as in importgit,
the path provided by git rev-parse --git-dir can be either a
relative or absolute path depending on the work tree.  Rather
than trying to deal with that, just use git -C.

Differential Revision:	https://reviews.freebsd.org/D8501
Reviewed by: markj


git-svn-id: https://svn.freebsd.org/base/head@308562 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f