MFC r302474 (By gnn)
On FreeBSD there is a setsockopt option SO_USER_COOKIE which allows
setting a 32 bit value on each socket. This can be used by applications
and DTrace as a rendezvous point so that an applicaton's data can
more easily be captured at run time. Expose the user cookie via
DTrace by updating the translator in tcp.d and add a quick test
program, a TCP server, that sets the cookie on each connection
accepted.

Sponsored by:	    Limelight Networks


git-svn-id: https://svn.freebsd.org/base/stable/11@308701 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f